+++ include/index/index.h
@@ -31,7 +31,7 @@ class Index {
 
   virtual auto insert(uint64_t id) -> bool = 0;
   virtual auto erase(uint64_t id) -> bool = 0;
-  virtual auto query(std::unique_ptr<VectorRecord> &record, int k) -> std::vector<int32_t> = 0;
+  virtual auto query(std::unique_ptr<VectorRecord> &record, int k) -> std::vector<uint64_t> = 0;
 };
 
 class GlobalIndex final : public Index {
