+++ include/index/knn.h
@@ -6,6 +6,6 @@ class Knn final : public Index {
   ~Knn() override;
   auto insert(uint64_t id) -> bool override;
   auto erase(uint64_t id) -> bool override;
-  auto query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<int32_t> override;
+  auto query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<uint64_t> override;
 };
 }  // namespace candy
