+++ include/index/vectraflow.h
@@ -8,6 +8,6 @@ class VectraFlow final : public Index {
   ~VectraFlow() override;
   auto insert(uint64_t id) -> bool override;
   auto erase(uint64_t id) -> bool override;
-  auto query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<int32_t> override;
+  auto query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<uint64_t> override;
 };
 }  // namespace candy
