+++ src/index/knn.cpp
@@ -9,7 +9,7 @@ auto candy::Knn::insert(uint64_t id) -> bool { return true; }
 
 auto candy::Knn::erase(uint64_t id) -> bool { return true; }
 
-auto candy::Knn::query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<int32_t> {
+auto candy::Knn::query(std::unique_ptr<VectorRecord>& record, int k) -> std::vector<uint64_t> {
   auto idxes = storage_manager_->topk(record, k);
   return idxes;
 }
